+++ NEWS.md
@@ -18,6 +18,7 @@
 - Update IRIs with 'marc' to use 'library' instead [MODLD-792](https://folio-org.atlassian.net/browse/MODLD-792)
 - Add Hub type and relation predicate [MODLD-870](https://folio-org.atlassian.net/browse/MODLD-870)
 - ResourceViewDeserializer introduced [MODLDI-5](https://folio-org.atlassian.net/browse/MODLDI-5)
+- Resource model made serializable to JSON [MODLDI-6](https://folio-org.atlassian.net/browse/MODLDI-6)
 
 ## 1.0.1 (03-12-2025)
 - Initial release